requirements.txt
requests
pytest
pyyaml
pydantic
psycopg2-binary
python-dotenv